+++ src/actions/mainApp.js
@@ -1,0 +1,12 @@
+export const SELECT_GAME = 'SELECT_GAME';
+
+
+/// main app
+
+export function selectGame(name) {
+  return {
+    type: SELECT_GAME,
+    payload: name
+  }
+}
+
