+++ src/store/bowling.js
@@ -1,0 +1,3 @@
+import {emptyFrames} from '../components/BowlingGame/utils';
+
+export const init = () => ({ frames: emptyFrames()});
