+++ src/store/index.js
@@ -1,11 +1,9 @@
-import {createCards} from '../components/MemoryGame/utils';
-import {emptyFrames} from '../components/BowlingGame/utils';
+import * as  memoryGame from './memory';
+import * as bowlingGame from './bowling';
 import * as appConstants from '../constants';
 
 export const initialState = {
     selectedGame: appConstants.MEMORY_GAME,
-    showAll:false,
-    cards: createCards(),
-    frames: emptyFrames(),
-    memoryClicks: 0
+    ...memoryGame.init(),
+    ...bowlingGame.init()
   };
