+++ src/store/memory.js
@@ -1,0 +1,3 @@
+import {createCards} from '../components/MemoryGame/utils';
+
+export const init = () => ({ showAll:false, cards: createCards(), memoryClicks: 0});
